requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
arrow==0.17.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
classifier==2.0
click==8.0.1
collections-extended==1.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.5
django-active-link==0.1.7
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
efficientnet==1.1.1
entrypoints==0.3
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imutils==0.5.4
install==1.3.4
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
MouseInfo==0.1.3
mysqlclient==2.0.1
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
nltk==3.6.2
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.7.1
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
pyrsistent==0.17.3
PySastrawi==1.2.0
PyScreeze==0.1.26
python-dateutil==2.8.1
python-decouple==3.4
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
Sastrawi==1.0.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.2
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.2
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.4
twilio==6.60.0
Unipath==1.1
urllib3==1.25.10
wcwidth==0.2.5
webdriver-manager==3.3.0
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipfile38==0.0.3
